fix(api): Cast int to long for CacheStats type compatibility
## Fix
- Cast totalHits and totalMisses to long in CacheStats constructor
- Cast HitCount and MissCount to long in TopQuestion mapping

## Build Errors Resolved
- CS0029: Cannot convert int to long (4 errors) - FIXED

Related: #423

+++ apps/api/src/Api/Services/AiResponseCacheService.cs
@@ -269,8 +269,8 @@ return total";
                 .Select(s => new TopQuestion
                 {
                     QuestionHash = s.QuestionHash,
-                    HitCount = s.HitCount,
-                    MissCount = s.MissCount,
+                    HitCount = (long)s.HitCount,
+                    MissCount = (long)s.MissCount,
                     LastHitAt = s.LastHitAt
                 })
                 .ToList();
@@ -309,8 +309,8 @@ return total";
 
             return new CacheStats
             {
-                TotalHits = totalHits,
-                TotalMisses = totalMisses,
+                TotalHits = (long)totalHits,
+                TotalMisses = (long)totalMisses,
                 HitRate = hitRate,
                 TotalKeys = totalKeys,
                 CacheSizeBytes = cacheSize,
